requirements.txt
openai
chromadb
python-dotenv
discord.py